tmux: move monitorWindowSize to platform-specific impls
